Move GetLoggingConfig from cmd/broker to pkg/utils
The GetLoggingConfig helper lived in package broker (cmd/broker/config.go)
even though it is generic and was being imported by non-broker commands
(jobsink, auth_proxy) just to read the logging ConfigMap. Move it into
pkg/utils alongside the other small cross-cutting helpers, update all
call sites, and delete the now-empty cmd/broker package.

Fixes #8712
